OperacionesMatematicas

+++ Extacion de mio/clase12/OperacionesMatematicas.java
@@ -33,7 +33,7 @@ public class OperacionesMatematicas {
         if(b == 0){
             total = 0.0;
         }else{
-            total = a / b;
+            total = (double) a / b;
         }
         return total;
     }
